+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ###### Breaking changes
 - Relation links managed by `DataStore` are now non-enumerable by default
+- Before calculating `Record#changes`, `toJSON` is called on if available.
 
 ##### 3.0.0-beta.3 - 27 April 2016